app/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2       
click==8.1.7
colorama==0.4.6
Flask==3.0.2
gcloud==0.18.3
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jws==0.1.3
MarkupSafe==2.1.5
oauth2client==4.1.3
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-jwt==2.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.2